(#5218) - add pouchdb-merge and pouchdb-adapter-utils
