+++ packages/pouchdb-adapter-leveldb-core/package.json
@@ -19,8 +19,10 @@
     "levelup": "1.3.2",
     "pouchdb-binary-utils": "5.4.0-prerelease",
     "pouchdb-collections": "1.0.1",
+    "pouchdb-adapter-utils": "5.4.0-prerelease",
     "pouchdb-errors": "5.4.0-prerelease",
     "pouchdb-json": "5.4.0-prerelease",
+    "pouchdb-merge": "5.4.0-prerelease",
     "pouchdb-md5": "5.4.0-prerelease",
     "pouchdb-promise": "5.4.0-prerelease",
     "pouchdb-utils": "5.4.0-prerelease",
