+++ packages/pouchdb-adapter-utils/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "pouchdb-adapter-indexeddb",
+  "name": "pouchdb-adapter-utils",
   "version": "5.4.0-prerelease",
-  "description": "PouchDB adapter using IndexedDB as its data store.",
+  "description": "Utilities for PouchDB adapters.",
   "main": "./lib/index.js",
   "keywords": [],
   "author": "Dale Harvey <dale@arandomurl.com>",
@@ -13,14 +13,11 @@
     "src"
   ],
   "dependencies": {
-    "js-extend": "1.0.1",
     "pouchdb-binary-utils": "5.4.0-prerelease",
     "pouchdb-collections": "1.0.1",
-    "pouchdb-adapter-utils": "5.4.0-prerelease",
     "pouchdb-errors": "5.4.0-prerelease",
-    "pouchdb-json": "5.4.0-prerelease",
     "pouchdb-merge": "5.4.0-prerelease",
-    "pouchdb-promise": "5.4.0-prerelease",
+    "pouchdb-md5": "5.4.0-prerelease",
     "pouchdb-utils": "5.4.0-prerelease"
   }
 }
