+++ packages/pouchdb-adapter-utils/src/index.js
@@ -1,0 +1,22 @@
+import parseDoc from './parseDoc';
+import {
+  invalidIdError,
+  normalizeDdocFunctionName,
+  parseDdocFunctionName
+} from 'pouchdb-utils';
+import { isDeleted , isLocalId } from 'pouchdb-merge';
+import preprocessAttachments from './preprocessAttachments';
+import processDocs from './processDocs';
+import updateDoc from './updateDoc';
+
+export {
+  invalidIdError,
+  isDeleted,
+  isLocalId,
+  normalizeDdocFunctionName,
+  parseDdocFunctionName,
+  parseDoc,
+  preprocessAttachments,
+  processDocs,
+  updateDoc
+};
