+++ packages/pouchdb-adapter-utils/src/preprocessAttachments.js
