+++ packages/pouchdb-adapter-websql-core/package.json
@@ -16,8 +16,10 @@
     "js-extend": "1.0.1",
     "pouchdb-binary-utils": "5.4.0-prerelease",
     "pouchdb-collections": "1.0.1",
+    "pouchdb-adapter-utils": "5.4.0-prerelease",
     "pouchdb-errors": "5.4.0-prerelease",
     "pouchdb-json": "5.4.0-prerelease",
+    "pouchdb-merge": "5.4.0-prerelease",
     "pouchdb-promise": "5.4.0-prerelease",
     "pouchdb-utils": "5.4.0-prerelease"
   }
