+++ packages/pouchdb-core/package.json
@@ -20,6 +20,7 @@
     "js-extend": "1.0.1",
     "pouchdb-collections": "1.0.1",
     "pouchdb-errors": "5.4.0-prerelease",
+    "pouchdb-merge": "5.4.0-prerelease",
     "pouchdb-promise": "5.4.0-prerelease",
     "pouchdb-utils": "5.4.0-prerelease",
     "scope-eval": "0.0.3"
