+++ packages/pouchdb-for-coverage/package.json
@@ -27,9 +27,11 @@
     "pouchdb-checkpointer": "5.4.0-prerelease",
     "pouchdb-collate": "1.2.0",
     "pouchdb-collections": "1.0.1",
+    "pouchdb-adapter-utils": "5.4.0-prerelease",
     "pouchdb-errors": "5.4.0-prerelease",
     "pouchdb-generate-replication-id": "5.4.0-prerelease",
     "pouchdb-mapreduce-utils": "5.4.0-prerelease",
+    "pouchdb-merge": "5.4.0-prerelease",
     "pouchdb-node": "5.4.0-prerelease",
     "pouchdb-promise": "5.4.0-prerelease",
     "pouchdb-utils": "5.4.0-prerelease",
