+++ packages/pouchdb-merge/README.md
@@ -1,0 +1,26 @@
+pouchdb-merge ![semver non-compliant](https://img.shields.io/badge/semver-non--compliant-red.svg)
+======
+
+PouchDB's CouchDB-style document merge algorithm as a standalone module.
+
+### Usage
+
+```bash
+npm install --save-exact pouchdb-merge
+```
+
+For full API documentation and guides on PouchDB, see [PouchDB.com](http://pouchdb.com/). For details on PouchDB sub-packages, see the [Custom Builds documentation](http://pouchdb.com/custom.html).
+
+
+### Warning: semver-free zone!
+
+This package is conceptually an internal API used by PouchDB or its plugins. It does not follow semantic versioning (semver), and rather its version is pegged to PouchDB's. Use exact versions when installing, e.g. with `--save-exact`.
+
+
+### Source
+
+PouchDB and its sub-packages are distributed as a [monorepo](https://github.com/babel/babel/blob/master/doc/design/monorepo.md).
+
+For a full list of packages, see [the GitHub source](https://github.com/pouchdb/pouchdb/tree/master/packages).
+
+
