+++ packages/pouchdb-merge/src/collectConflicts.js
