+++ packages/pouchdb-merge/src/collectLeaves.js
