+++ packages/pouchdb-merge/src/compactTree.js
