+++ packages/pouchdb-merge/src/index.js
@@ -1,0 +1,23 @@
+import collectConflicts from './collectConflicts';
+import collectLeaves from './collectLeaves';
+import compactTree from './compactTree';
+import merge from './merge';
+import revExists from './revExists';
+import rootToLeaf from './rootToLeaf';
+import traverseRevTree from './traverseRevTree';
+import winningRev from './winningRev';
+import isDeleted from './isDeleted';
+import isLocalId from './isLocalId';
+
+export {
+  collectConflicts,
+  collectLeaves,
+  compactTree,
+  isDeleted,
+  isLocalId,
+  merge,
+  revExists,
+  rootToLeaf,
+  traverseRevTree,
+  winningRev
+};
