+++ packages/pouchdb-merge/src/isLocalId.js
