+++ packages/pouchdb-merge/src/revExists.js
