+++ packages/pouchdb-merge/src/rootToLeaf.js
