+++ packages/pouchdb-merge/src/traverseRevTree.js
