+++ packages/pouchdb-merge/src/winningRev.js
