+++ packages/pouchdb-utils/src/normalizeDdocFunctionName.js
