+++ packages/pouchdb-utils/src/parseDdocFunctionName.js
